requirements.txt
requests
python-dotenv
beautifulsoup4
sentence-transformers
scikit-learn
numpy
flask
pandas